z80: fix strlen to be size_t properly

+++ Kernel/cpu-z80/cpu.h
@@ -38,7 +38,7 @@ extern void *memmove(void *dest, const void *src, size_t n);
 extern int memcmp(const void * a, const void *b, size_t n);
 extern int strcmp(const char *, const char *);
 
-extern int16_t strlen(const char *p);
+extern size_t strlen(const char *p);
 
 #define	staticfast	static/* User's structure for times() system call */
 typedef unsigned long clock_t;
