+++ Kernel/platform-kc87/main.c
@@ -70,9 +70,9 @@ void do_beep(void)
  *	the runtime
  */
 
-int strlen(const char *p)
+size_t strlen(const char *p)
 {
-	int len = 0;
+	size_t len = 0;
 	while(*p++)
 		len++;
 	return len;
