+++ Kernel/platform-zx+3/main.c
@@ -35,9 +35,9 @@ void plt_interrupt(void)
  *	the runtime
  */
 
-int strlen(const char *p)
+size_t strlen(const char *p)
 {
-  int len = 0;
+  size_t len = 0;
   while(*p++)
     len++;
   return len;
