+++ Kernel/platform-zxdiv/main.c
@@ -34,9 +34,9 @@ void plt_interrupt(void)
  *	the runtime
  */
 
-int strlen(const char *p)
+size_t strlen(const char *p)
 {
-  int len = 0;
+  size_t len = 0;
   while(*p++)
     len++;
   return len;
